connect product detail view to core data and implement product detail next and previous navigation

+++ A2_iOS_penny_101488485/ProductRepository.swift
@@ -63,7 +63,7 @@ final class ProductRepository {
         product.productPrice = NSDecimalNumber(decimal: price)
         product.productProvider = provider
         product.productCategory = category
-        product.productStockQty = stockQty
+        product.productStockQty = Int32(stockQty)
 
         return saveContext()
     }
